==> rscript/01_TimeGPT.ipynb <==
!pip install -Uqq nixtla

# Load libraries

from nixtla import NixtlaClient
import pandas as pd
import numpy as np

# initiate Nixtla client

nixtla_client = NixtlaClient(
    api_key = 'nixtla-tok-fQ023pkMcbq0hsxZCUNUyHruTjNHOe07YkG6PIdlUduZaQeyXG18s1e59Fjt2FieAZBWOPxBKTgEeJeD'
)

# key validation

nixtla_client.validate_api_key()

# Load data

from google.colab import files
uploaded = files.upload()

# read data

train_tscv_malaria = pd.read_csv("train_tscv_malaria.csv")
test_tscv_malaria = pd.read_csv("test_tscv_malaria.csv")

train_tscv_no_malaria = pd.read_csv("train_tscv_no_malaria.csv")
test_tscv_no_malaria = pd.read_csv("test_tscv_no_malaria.csv")

# make the month column as date format

train_tscv_malaria['month'] = pd.to_datetime(train_tscv_malaria['month']) + pd.offsets.MonthEnd(0)
train_tscv_malaria = train_tscv_malaria.sort_values(by = 'month')

test_tscv_malaria['month'] = pd.to_datetime(test_tscv_malaria['month']) + pd.offsets.MonthEnd(0)
test_tscv_malaria = test_tscv_malaria.sort_values(by = 'month')
test_tscv_malaria = test_tscv_malaria.drop('quantity', axis=1)

train_tscv_no_malaria['month'] = pd.to_datetime(train_tscv_no_malaria['month']) + pd.offsets.MonthEnd(0)
train_tscv_malaria = train_tscv_malaria.sort_values(by = 'month')

test_tscv_no_malaria['month'] = pd.to_datetime(test_tscv_no_malaria['month']) + pd.offsets.MonthEnd(0)
test_tscv_no_malaria = test_tscv_no_malaria.sort_values(by = 'month')
test_tscv_no_malaria = test_tscv_no_malaria.drop('quantity', axis=1)

train_tscv_malaria = train_tscv_malaria.rename(columns={
    'month': 'ds',
    'quantity': 'y'
})

test_tscv_malaria = test_tscv_malaria.rename(columns={
    'month': 'ds',
    'quantity': 'y'
})

train_tscv_no_malaria = train_tscv_no_malaria.rename(columns={
    'month': 'ds',
    'quantity': 'y'
})

test_tscv_no_malaria = test_tscv_no_malaria.rename(columns={
    'month': 'ds',
    'quantity': 'y'
})

train_tscv_malaria.head()

# define parameters

horizon = 6

# Define prediction intervals
levels = np.arange(0.01, 1.00, 0.01)

# model fit with regressors

# products with malaria seasonality
fc_malaria_reg = nixtla_client.forecast(
    df=train_tscv_malaria,
    h=horizon,
    quantiles=levels,
    X_df=test_tscv_malaria,
    freq = 'M',
)

# products without malaria seasonality
fc_no_malaria_reg = nixtla_client.forecast(
    df=train_tscv_no_malaria,
    h=horizon,
    quantiles=levels,
    X_df=test_tscv_no_malaria,
    freq = 'M',
)

# model fit without regressors

# products with malaria seasonality
fc_malaria = nixtla_client.forecast(
    df=train_tscv_malaria,
    h=horizon,
    quantiles=levels,
    #X_df=test_tscv_malaria,
    freq = 'M',
)

# products without malaria seasonality
fc_no_malaria = nixtla_client.forecast(
    df=train_tscv_no_malaria,
    h=horizon,
    quantiles=levels,
    #X_df=test_tscv_no_malaria,
    freq = 'M',
)

# save files

fc_malaria_reg.to_csv('timegpt_malaria_reg.csv', index = False)
files.download("timegpt_malaria_reg.csv")

fc_no_malaria_reg.to_csv('timegpt_no_malaria_reg.csv', index = False)
files.download("timegpt_no_malaria_reg.csv")

fc_malaria.to_csv('timegpt_malaria.csv', index = False)
files.download("timegpt_malaria.csv")

fc_no_malaria.to_csv('timegpt_no_malaria.csv', index = False)
files.download("timegpt_no_malaria.csv")
